Discord処理を外だし
